requirement.txt
sklearn == 0.23.2
numpy == 1.19.3
pandas == 1.1.3
matplotlib == 3.3.2
dlib == 19.21.0
imutils == 0.5.3
natsort == 7.0.1
cv2 == 4.5.0
seaborn == 0.11.0
pickleshare == 0.7.5
tensorflow-determinism == 0.3.0
tensorflow-estimator == 2.0.1
tensorflow-gpu == 2.0.0